change class method, change fn on-submit-query, add preloader on index.html
