+++ src/js/serviceApiFilmTrend.js
@@ -16,7 +16,7 @@ export default class FilmApiTrendFetch {
 
   async fetchFilmsGenres() {
     return await fetch(
-      `${GENRES_URL}?api_key=${API_KEY}&language=${this.currentLang}`
+      `${GENRES_URL}?api_key=${API_KEY}&language=${this.currentLang}page=${this.page}`
     )
       .then(res => res.json())
       .then(data => {
@@ -70,6 +70,7 @@ export default class FilmApiTrendFetch {
       console.log(error);
     }
   }
+
   async fetchSearchFilms() {
     return await fetch(
       `${SEARCH_FILMS_URL}?api_key=${API_KEY}&language=${this.currentLang}&query=${this.query}`
@@ -83,17 +84,68 @@ export default class FilmApiTrendFetch {
       .catch(err => console.log(err));
   }
 
+  async searchFilmsAndGenres() {
+    try {
+      const response = await this.fetchSearchFilms();
+      // console.log(response);
+      const data = {
+        films: this.films,
+        total_results: response.total_results,
+      };
+      // console.log(data.films);
+      const genresIdArrFromApi = this.genres;
+      // console.log(genresIdArrFromApi);
+
+      for (let film of data.films) {
+        film.genre_ids = searchGenres(film.genre_ids);
+        // console.log(film.genre_ids);
+        // форматуємо рейтинг
+        film.vote_average = film.vote_average.toFixed(1);
+        // форматуємо дату виходу фільму
+        film.release_date = film.release_date.slice(0, 4);
+        // форматуємо кількість жанрів фільму
+        if (film.genre_ids.length === 0) {
+          film.genre_ids[0] = 'no movie genre';
+        }
+        if (film.genre_ids.length >= 3) {
+          film.genre_ids[2] = 'Other';
+        }
+        film.genre_ids = film.genre_ids.slice(0, 3).join(', ');
+      }
+
+      function searchGenres(ids) {
+        let genresNamesArr = [];
+        let searchId = ids;
+        let genreName;
+
+        for (let i = 0; i < ids.length; i += 1) {
+          genreName = genresIdArrFromApi.find(
+            list => list.id === searchId[i]
+          ).name;
+          genresNamesArr.push(genreName);
+        }
+        return genresNamesArr;
+      }
+      return data;
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
   async fetchFilmCard() {
     // this.movie_id = 766475;
-    try{      
+    try {
       return await fetch(
-      `${CARD_MOVIE}${this.movie_id}?api_key=${API_KEY}&language=${this.currentLang}`)
-      .then(res => res.json())
-      .then(data => {
-        console.log(data)  
-        console.log(data.overview); 
-      })}
-      catch(err) {console.log(err)};      
+        `${CARD_MOVIE}${this.movie_id}?api_key=${API_KEY}&language=${this.currentLang}`
+      )
+        .then(res => res.json())
+        .then(data => {
+          console.log(data);
+          console.log(data.overview);
+        });
+    } catch (err) {
+      console.log(err);
+    }
   }
 
   get lang() {
@@ -110,10 +162,10 @@ export default class FilmApiTrendFetch {
     this.page = value;
   }
 
-  get idFilm (){
+  get idFilm() {
     return this.movie_id;
   }
-  set idFilm (value) {
+  set idFilm(value) {
     this.movie_id = value;
   }
 }
